fix(compose): the query decode was strict below a lenient entry point
Findings from the Fable review pass, all six applied.

THE ONE THAT MATTERED. The reader called r.URL.Query(), which runs ParseQuery
and DISCARDS its error — keeping the pairs that parsed and dropping the ones
that did not. Every strict rule below it was therefore applied to a value set
the caller had not sent:

  - ?payload=p&exact=%zz  the optional argument vanished and the call succeeded,
    which is the silent loss the unknown_parameter refusal exists to prevent;
  - ?payload=a&payload=%zz  arrived single-valued, so the repeated-key refusal
    never fired and the seam picked a winner after all;
  - ?a=1;b=2  dropped the WHOLE query, so a route with no required arguments
    answered 200 having read none of them.

Now url.ParseQuery on RawQuery, refused as malformed_query. The docstring said
"It is STRICT, and it is the only thing that is"; it is now true of the entry
point as well as the body.

Also:

  - `number` is held to the contract's grammar, not strconv's. ParseFloat
    accepts NaN, Inf, Infinity, hex floats (0x1p10 → 1024) and underscored
    digits — none of which the published schema admits. The first three were
    refused only by json.Marshal FAILING, which handed the caller Go's own
    error text as the validation detail; the rest were served. decimalNumeral
    also puts `number` and `integer` on one grammar, since base-10 ParseInt
    already refused underscores and hex.
  - the refusal names the same argument every time. decode returns on the first
    bad one and its input is a map, so two bad arguments named one on this call
    and the other on the next.
  - path-item-level `parameters` are refused. OpenAPI applies them to every
    operation beneath, this reader sees only an operation's own, so they were
    published to every client and read by nothing — a GET declaring `limit`
    there generated a route that refuses ?limit=5 forever. Exactly the fault
    argumentSchema names one level down, and harmless only until GET became
    declarable.
  - validateQueryEncodable now checks `required` ⊆ properties, and that
    `required` is a list. Two comments claimed Validate already refused
    everything the mount could trip on; it did not, and since mounting PANICS,
    a declaration Validate blessed could be a boot crash instead of a named
    generation failure.
  - the notes fragment's signature comment claimed a query payload reaches the
    access log. It does not: AccessLog records method, path and status, never
    RawQuery. Browser history, Referer and intermediary proxies are the real
    reasons, the request-line cap is the independent one, and the corrected
    comment says outright that query arguments do NOT reach an operator's log —
    the previous text taught authors the opposite.

Two comment corrections with no behaviour change: HEAD may not be DECLARED but
is still served through every GET pattern by Go's ServeMux, and the bodyless
routes' arguments are bounded by http.DefaultMaxHeaderBytes rather than by the
explicit body cap.

extroutes.go passed the 500-LOC cap again, so the argument-reading half moved to
extrouteargs.go — which is the honest seam: that file is the whole of the
argument validation in the system, and extroutes.go owns mounting and the
response.

Refs #993
